Added markdown stuff
Feel free to add to this - the background information may be outright
incorrect. I also deleted the stuff about the dash in the “basic
markdown” section about its use in the state’s content element - as far
as I can tell it doesn’t create an unordered list in Odyssey as it
would in regular markdown. I also reformatted that section, feel free
to ignore those changes!